workflow API + eclipse files

+++ grails-app/domain/org/chai/kevin/security/SurveyUser.groovy
@@ -4,7 +4,7 @@ class SurveyUser extends User {
 
 	// TODO assign several surveys to a user
 	// instead of an organisation unit
-	int entityId;
+	Long entityId;
 
 	static constraints = {
 		entityId (nullable: false)
